python_backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.1.1

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain and LangGraph (Latest versions)
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-openai==0.2.14
langgraph==0.2.60

# Vector Database - Milvus
pymilvus==2.5.0

# Graph Database - Neo4j
neo4j==5.27.0

# Search Engine - Elasticsearch
elasticsearch==8.17.0

# Redis
redis==5.0.1

# Async support
httpx==0.28.1
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0

# SSE for streaming
sse-starlette==2.2.1
